backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
python-docx==0.8.11
PyPDF2==3.0.1
flask-cors==3.0.10
nltk==3.8.1
python-dotenv==0.19.0
pytest==7.0.1
pytest-cov==3.0.0